- Rework `Show-ADTHelpConsole` to display help for extension modules also.
This required more form work than I'd have liked. Where do I begin... Firstly, the ComboBox's default border colour doesn't match that of the Panels, so attempted to sit it inside a Panel, however the control's size varies depending on DPI and the Panel even on AutoSize would not AutoSize correctly.

When having the ComboBox outside of the Panel, getting its sizing to match that of the ListBox Panel across all DPIs was next to impossible. As such, move the ListBox out of a Panel, it was only in it to give it a consistently coloured border with the Panel for the RichTextBox. Once that was done, having the Panel-less ListBox line up with the RichTextBox's Panel was not working either. Sooo, we're back to a Fixed3D RichTextBox to have everything line up...

Finalised the lining up of everything by using a consistent font so the AutoScaling works 100%, and by interrogating the display's DPI and using its scale compared to 96 DPI to adjust the height/width of the overall form by its ratio. This has been tested at 100%, 125% and 150% and renders well.