- Rework process running test in RunningProcessService again.
The setup I put in just prior worked great for elevated processes we own, but fails for processes running as another user. As such, just try to get the process's command line. If we can get it, the process is running. If we get an ArgumentException saying "Value does not fall within the expected range.", we know the process Id is no more.