- Add config/ADMX comparison code to the module build process.
This also refactors some newly added config options. We can hook up `DialogAccentColor` to the WinForms stuff (I think). `ExpiryDuration` is a duplicate of `DefaultTimeout` so it can go. The remaining two can't be config options for both dialogs because they will need to vary for each dialog unless specified. I'd prefer both of these to just be arguments on the dialog functions.