- Add `CountdownStopwatch` to `CloseAppsDialogOptions` class.
We're going to need this so that if the dialog is closed while processes are running and needs to re-open, the countdown knows where things are up to.

This is better than other options such as storing the current date/time + the countdown duration as this doesn't account for the time it takes to evaluate running processes, instantiate a dialog, etc. The current fluent dialog code will need its countdown code refactored to use this.