What? Me? Nothing, just chillin and not breaking stuff, no sir....
Migrates the Wpf.Ui to iNKORE's ModernWPF UI library for improved look and feel as well as better maintainability.

This includes updating the project file, XAML namespaces, and code-behind to align with the new library's structure and components.
Removes the dependency on the older Wpf.Ui library and updates the theming and styling to match ModernWPF's design principles.