Change switch `-AllowRebootPassThru` to `-SuppressRebootPassThru`.
The way this was isn't the most ideal as it's natural to want to allow a reboot code to go back through to your deployment mechanism. Ideally this switch shouldn't exist at all, but we supply the antonym for those who were relying on it. The right solution is to just control it in SCCM or Intune on a per-app basis.